+++ src/SelectableColumnTypes.ts
@@ -1,0 +1,1 @@
+export type SelectableColumnTypes = string | number | boolean | Date | undefined | null;